pyre-jit-trace: extract arith/record helpers to arith.rs
Move the int_*/float_*/ptr_* kind-parameterized record helpers
(binop/unop recorders, pure-binop eval, ref/ptr guards, cast) into
arith.rs.
